# internal/domain/house.go
package domain

import (
	"time"
)

// House represents a boarding house entity.
type House struct {
	ID        int64     `json:"id"`
	Name      string    `json:"name"`
	Address   string    `json:"address"`
	CreatedAt time.Time `json:"created_at"`
	UpdatedAt *time.Time `json:"updated_at"`
}