# openapi.yaml
openapi: 3.0.3
info:
  title: Tro-Go API Documentation
  description: |
    Tài liệu API cho dự án Tro-Go (Quản lý nhà trọ và khách thuê).
    Dành cho Frontend Team sử dụng ReDoc để render.
  version: 1.0.0

servers:
  - url: http://localhost:8080/api/v1
    description: Local development server

tags:
  - name: Auth
    description: Đăng ký, đăng nhập và thông tin tài khoản.
  - name: Houses
    description: Quản lý thông tin các tòa nhà/nhà trọ.

paths:
  /auth/register:
    post:
      tags:
        - Auth
      summary: Đăng ký tài khoản mới
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          description: Đăng ký thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseUser'
        '400':
          $ref: '#/components/responses/BadRequestError'

  /auth/login:
    post:
      tags:
        - Auth
      summary: Đăng nhập
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Đăng nhập thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseToken'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  /auth/me:
    get:
      tags:
        - Auth
      summary: Lấy thông tin tài khoản hiện tại
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseUser'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  /houses:
    get:
      tags:
        - Houses
      summary: Danh sách nhà trọ (phân trang)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseHouseList'
    post:
      tags:
        - Houses
      summary: Tạo nhà trọ mới
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/House'
      responses:
        '201':
          description: Tạo thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseHouse'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /houses/{id}:
    get:
      tags:
        - Houses
      summary: Chi tiết nhà trọ theo ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseHouse'
        '404':
          $ref: '#/components/responses/NotFoundError'
    put:
      tags:
        - Houses
      summary: Cập nhật thông tin nhà trọ
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/House'
      responses:
        '200':
          description: Cập nhật thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseMessage'
        '404':
          $ref: '#/components/responses/NotFoundError'
    delete:
      tags:
        - Houses
      summary: Xóa nhà trọ
      description: Yêu cầu quyền `house:delete`
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Xóa thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseMessage'
        '404':
          $ref: '#/components/responses/NotFoundError'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    # --- Generic Response ---
    Meta:
      type: object
      properties:
        total:
          type: integer
          example: 50
        current_page:
          type: integer
          example: 1
        last_page:
          type: integer
          example: 5
        limit:
          type: integer
          example: 10

    # --- Concrete Responses ---
    ApiResponseUser:
      type: object
      properties:
        status:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/User'

    ApiResponseToken:
      type: object
      properties:
        status:
          type: boolean
          example: true
        data:
          type: object
          properties:
            token:
              type: string
              example: "eyJhbGciOiJIUzI1..."

    ApiResponseHouse:
      type: object
      properties:
        status:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/House'

    ApiResponseHouseList:
      type: object
      properties:
        status:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/House'
        meta:
          $ref: '#/components/schemas/Meta'

    ApiResponseMessage:
      type: object
      properties:
        status:
          type: boolean
          example: true
        data:
          type: string
          example: "success message"

    ApiResponseError:
      type: object
      properties:
        status:
          type: boolean
          example: false
        data:
          type: string
          example: "error message details"

    # --- Entities ---
    User:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        username:
          type: string
          example: "johndoe"
        email:
          type: string
          format: email
          example: "john@example.com"
        password:
          type: string
          format: password
          writeOnly: true
        full_name:
          type: string
          example: "John Doe"
        role:
          type: string
          example: "host"
        permissions:
          type: array
          items:
            type: string
          example: ["house:create", "house:delete"]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          example: "johndoe"
        password:
          type: string
          format: password
          example: "secret123"

    House:
      type: object
      required: [name, address]
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          example: "Nhà trọ Hạnh Phúc"
        address:
          type: string
          example: "123 Đường ABC, Quận 1, TP.HCM"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

  responses:
    BadRequestError:
      description: Dữ liệu gửi lên không hợp lệ.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponseError'
    UnauthorizedError:
      description: Token không hợp lệ hoặc không có quyền truy cập.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponseError'
          example:
            status: false
            data: "Unauthorized"
    NotFoundError:
      description: Tài nguyên không tồn tại.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponseError'
    InternalServerError:
      description: Lỗi hệ thống.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponseError'